% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{Gibbs_a}
\alias{Gibbs_a}
\title{Sample posterior for hyperparameter 'a'}
\usage{
Gibbs_a(nu, a, s_a = 1, r_a = 1)
}
\arguments{
\item{nu}{p*1 vector of local shrinkage parameters}

\item{s_a}{scalar, shape parameter for prior on 'a'}

\item{r_a}{scalar, rate parameter for prior on 'a'}

\item{a_current}{scalar, current value of 'a'}
}
\value{
scalar a_new
}
\description{
Sample posterior for hyperparameter 'a'
}
