% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{Gibbs_b}
\alias{Gibbs_b}
\title{Sample posterior for hyperparameter 'b'}
\usage{
Gibbs_b(lambda, b, s_b = 1, r_b = 1)
}
\arguments{
\item{lambda}{p*1 vector of local shrinkage parameters}

\item{b_current}{scalar, current value of 'b'}

\item{s_b_prior}{scalar, shape parameter for prior on 'b' (renamed to avoid conflict with local b)}

\item{r_b_prior}{scalar, rate parameter for prior on 'b'}
}
\value{
scalar b_new
}
\description{
Sample posterior for hyperparameter 'b'
}
