% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_beta.R
\name{Gibbs_beta}
\alias{Gibbs_beta}
\title{Draw posterior beta under the nu/lambda TPB augmentation}
\usage{
Gibbs_beta(
  X,
  y,
  a = NULL,
  b = NULL,
  phi = NULL,
  sigmaSq = NULL,
  nu = NULL,
  lambda = NULL,
  woodbury = FALSE,
  diagX = FALSE,
  omega = NULL
)
}
\arguments{
\item{X}{n*p design matrix}

\item{y}{n*1 response vector}

\item{phi}{Local beta variance scale. This is omega in the EB sampler,
and phi * a / b inside the fully Bayesian sampler.}

\item{sigmaSq}{error variance (scalar)}

\item{nu, lambda}{p*1 local-scale vectors}

\item{woodbury}{binary logic variable, apply Woodbury identity or not}

\item{diagX}{binary logic variable, if TRUE assumes X is diagonal}
}
\value{
posterior vector beta
}
\description{
Draw posterior beta under the nu/lambda TPB augmentation
}
