% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_nu_lambda.R
\name{Gibbs_lambda}
\alias{Gibbs_lambda}
\title{Sample lambda in the nu/lambda TPB augmentation}
\usage{
Gibbs_lambda(p, b, phi, beta, nu)
}
\arguments{
\item{p}{ncol(X)}

\item{b}{TPB shape parameter}

\item{phi}{Local beta variance scale, named phi for compatibility with
Gibbs_beta(). This is omega in the EB sampler and phi * a / b inside
the fully Bayesian sampler.}

\item{beta}{p*1 coefficient vector}

\item{nu}{p*1 local nu vector}
}
\value{
posterior vector lambda
}
\description{
Sample lambda in the nu/lambda TPB augmentation
}
