% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_zeta_psi.R
\name{Gibbs_psi}
\alias{Gibbs_psi}
\title{Sample local shrinkage ratio psi (IGIG Step 2)}
\usage{
Gibbs_psi(p, b, phi, beta, zeta)
}
\arguments{
\item{p}{ncol(X)}

\item{b}{hyperparameter b}

\item{phi}{global shrinkage parameter}

\item{beta}{p*1 coefficient vector}

\item{zeta}{p*1 vector of local auxiliary variables}
}
\description{
Sample local shrinkage ratio psi (IGIG Step 2)
}
