% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_sigma.R
\name{Gibbs_sigmaSq}
\alias{Gibbs_sigmaSq}
\title{Sample error variance sigmaSq}
\usage{
Gibbs_sigmaSq(n, X, y, beta)
}
\arguments{
\item{n}{Number of observations.}

\item{X}{n*p design matrix.}

\item{y}{n*1 response vector.}

\item{beta}{p*1 coefficient vector.}
}
\value{
A scalar draw of sigmaSq.
}
\description{
Sample error variance sigmaSq
}
