% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_nu_lambda.R
\name{Gibbs_xi}
\alias{Gibbs_xi}
\title{Sample xi in the nu/lambda TPB augmentation}
\usage{
Gibbs_xi(p, a, nu)
}
\arguments{
\item{p}{ncol(X)}

\item{a}{TPB shape parameter}

\item{nu}{p*1 local nu vector}
}
\value{
posterior vector xi
}
\description{
Sample xi in the nu/lambda TPB augmentation
}
