% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_zeta_psi.R
\name{Gibbs_zeta}
\alias{Gibbs_zeta}
\title{Sample local auxiliary variable zeta (IGIG Step 1)}
\usage{
Gibbs_zeta(p, a, b, psi)
}
\arguments{
\item{p}{ncol(X)}

\item{a}{hyperparameter a}

\item{b}{hyperparameter b}

\item{psi}{p*1 vector of local shrinkage ratios}
}
\description{
Sample local auxiliary variable zeta (IGIG Step 1)
}
