% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConverged_emp.R
\name{isConverged_hyper}
\alias{isConverged_hyper}
\title{define a convergence criterion}
\usage{
isConverged_hyper(vec1, vec2, vec3, delta1, delta2, window_size)
}
\arguments{
\item{vec1}{vector of a}

\item{vec2}{vector of b}

\item{vec3}{vector of omega}

\item{delta1}{numeric, generally fix delta1 at 0.001 to avoid devision by 0}

\item{delta2}{numeric}

\item{window_size}{numeric, size of the sliding window}
}
\description{
define a convergence criterion
}
