% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marginal_beta.R
\name{log_marginal_posterior}
\alias{log_marginal_posterior}
\title{Compute marginal log posterior for TPB hyperparameters}
\usage{
log_marginal_posterior(
  a,
  b,
  phi,
  beta_vec,
  s_prior_a,
  r_prior_a,
  s_prior_b,
  r_prior_b,
  scale_phi,
  prior_type_a = "gamma",
  prior_type_b = "gamma",
  scale_a = 1,
  scale_b = 1
)
}
\arguments{
\item{a}{shape parameter a}

\item{b}{shape parameter b}

\item{phi}{global shrinkage parameter}

\item{beta_vec}{Coefficient vector.}

\item{s_prior_a}{Prior shape parameter for a.}

\item{r_prior_a}{Prior rate parameter for a.}

\item{s_prior_b}{Prior shape parameter for b.}

\item{r_prior_b}{Prior rate parameter for b.}

\item{scale_phi}{Half-Cauchy scale parameter for phi.}

\item{prior_type_a}{prior type for a: "gamma" or "hcauchy"}

\item{prior_type_b}{prior type for b: "gamma" or "hcauchy"}

\item{scale_a}{Half-Cauchy scale parameter for a}

\item{scale_b}{Half-Cauchy scale parameter for b}
}
\description{
Compute marginal log posterior for TPB hyperparameters
}
