% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{M.step}
\alias{M.step}
\title{E-M steps}
\usage{
M.step(sample_matrix)
}
\arguments{
\item{sample_matrix}{a matrix with dimension num_sample*p - nu for estimating a and lambda for estimating b}
}
\description{
E-M steps
}
