% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_single.R
\name{run_marginal_mh_uni}
\alias{run_marginal_mh_uni}
\title{single variable Marginal MH}
\usage{
run_marginal_mh_uni(
  beta_vec,
  target_param,
  current_a,
  current_b,
  current_phi,
  s_prior_a,
  r_prior_a,
  s_prior_b,
  r_prior_b,
  scale_phi,
  prior_type_a = "gamma",
  prior_type_b = "gamma",
  scale_a = 1,
  scale_b = 1,
  mh_step = 0.1,
  max_log_mh_step = Inf
)
}
\arguments{
\item{target_param}{("a", "b", or "phi")}
}
\description{
single variable Marginal MH
}
