% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{run_mh_gamma}
\alias{run_mh_gamma}
\title{Metropolis-Hastings Algorithm for Gamma Prior with log-normal proposal}
\usage{
run_mh_gamma(current_val, x, s_prior, r_prior, w = 0.05)
}
\arguments{
\item{current_val}{(a or b)}

\item{x}{(nu or lambda)}

\item{w}{step size}
}
\description{
Metropolis-Hastings Algorithm for Gamma Prior with log-normal proposal
}
