% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{run_mh_hcauchy}
\alias{run_mh_hcauchy}
\title{Metropolis-Hastings Algorithm for Half-Cauchy Prior with log-normal proposal}
\usage{
run_mh_hcauchy(current_val, x, scale_prior, w = 0.05)
}
\arguments{
\item{current_val}{(a or b)}

\item{x}{(nu or lambda)}

\item{scale_prior}{Half-Cauchy(0, scale)}

\item{w}{step_size}
}
\description{
Metropolis-Hastings Algorithm for Half-Cauchy Prior with log-normal proposal
}
