% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{run_miller_algorithm}
\alias{run_miller_algorithm}
\title{Miller's Approximation Algorithm for Gamma Shape Posterior
Find the parameters A and B of a Gamma(A, B) distribution that approximates the posterior of the shape parameter 'a'.
Model: X ~ Gamma(a, a/mu) (implying mean = mu)
Prior: a ~ Gamma(s_prior, r_prior)}
\usage{
run_miller_algorithm(x, mu, s_prior, r_prior, tol = 1e-04, max_iter = 20)
}
\arguments{
\item{x}{Vector of data (e.g., nu for parameter a, or lambda for parameter b)}

\item{mu}{Current value of the parameter (a or b). In the rate=1 parameterization, mu = a (or b).}

\item{s_prior}{Prior shape parameter for a (or b)}

\item{r_prior}{Prior rate parameter for a (or b)}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
A list containing A and B
}
\description{
Miller's Approximation Algorithm for Gamma Shape Posterior
Find the parameters A and B of a Gamma(A, B) distribution that approximates the posterior of the shape parameter 'a'.
Model: X ~ Gamma(a, a/mu) (implying mean = mu)
Prior: a ~ Gamma(s_prior, r_prior)
}
