% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamples_full_fixed_ab.R
\name{tpb_fullyBayes_fixab}
\alias{tpb_fullyBayes_fixab}
\title{Fully Bayesian Gibbs Sampler for the TPB Normal Model with Fixed a and b}
\usage{
tpb_fullyBayes_fixab(
  X,
  y,
  a,
  b,
  num_burnin,
  num_samples,
  woodbury = TRUE,
  diagX = FALSE,
  fix_phi = FALSE,
  phi_val = NULL,
  scale_phi = NULL
)
}
\arguments{
\item{num_burnin}{The number of MCMC iterations to discard as burn-in.}

\item{num_samples}{The number of MCMC iterations to save after the burn-in period.}

\item{fix_phi}{Logical. If TRUE, phi is fixed to the provided phi_val without sampling.}

\item{phi_val}{Numeric. The fixed value for phi (if fix_phi = TRUE) or the initial starting value.}

\item{scale_phi}{Half-Cauchy scale for the typical phi value. Internally this
uses sqrt(phi) ~ C+(0, sqrt(scale_phi)) for a conjugate update.}
}
\value{
list of posterior samples

list of posterior samples
}
\description{
Fully Bayesian Gibbs Sampler for the TPB Normal Model with Fixed a and b
}
