% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullGibbs.R
\name{fullGibbs}
\alias{fullGibbs}
\title{Fully Gibbs Sampler Function with Miller's Approximation}
\usage{
fullGibbs(
  X,
  y,
  num_output = 10000,
  num_burnin = 10000,
  thin = 1,
  woodbury = TRUE,
  diagX = FALSE,
  proposal_type = "separate",
  mh_step_a = 0.1,
  mh_step_b = 0.1,
  mh_step_phi = 0.1,
  adapt_block_size = 100,
  r_opt = 0.3,
  cov_matrix = matrix(c(0.01, -0.005, -0.005, 0.01), 2, 2),
  hyper_params = list(prior_type_a = "gamma", prior_type_b = "gamma", s_a = 1.5, r_a = 1,
    s_b = 1.5, r_b = 1, scale_a = 1, scale_b = 1, scale_phi = 1)
)
}
\arguments{
\item{X}{n*p design matrix}

\item{y}{n*1 response vector}

\item{num_output}{Integer, number of posterior samples to save (after burn-in)}

\item{num_burnin}{Integer, number of burn-in iterations to discard}

\item{thin}{Integer, thinning interval for saved posterior samples}

\item{woodbury}{Logical, use Woodbury identity in beta update}

\item{diagX}{Logical, assume diagonal X}

\item{proposal_type}{String, "separate" or "bi_fixed" or "bi_adaptive" or "all_adaptive"}

\item{mh_step_a, }{mh_step_b, mh_step_phi step size for each hyper-parameter under the 'separate' mode}

\item{adapt_block_size}{Integer, number of iterations per adaptation block}

\item{r_opt}{Numeric, target acceptance rate for adaptive MH (default 0.3)}

\item{cov_matrix}{Matrix, 2x2 covariance matrix for bi_fixed proposals}

\item{hyper_params}{List of hyperparameters (prior_type_a, prior_type_b, s_a, r_a, s_b, r_b, scale_a, scale_b, scale_phi)}

\item{mh_step}{scalar, step size for MH uniform proposal}
}
\value{
A list containing posterior samples matrices, acceptance rates, covariance matrix.
}
\description{
Fully Gibbs Sampler Function with Miller's Approximation
}
