% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullGibbs.R
\name{fullGibbs}
\alias{fullGibbs}
\title{Fully Gibbs Sampler Function with Miller's Approximation}
\usage{
fullGibbs(
  X,
  y,
  num_output = 10000,
  num_burnin = 10000,
  woodbury = FALSE,
  diagX = FALSE,
  hyper_params = list(s_a = 1, r_a = 1, s_b = 1, r_b = 1)
)
}
\arguments{
\item{X}{n*p design matrix}

\item{y}{n*1 response vector}

\item{num_output}{Integer, number of posterior samples to save (after burn-in)}

\item{num_burnin}{Integer, number of burn-in iterations to discard}

\item{woodbury}{Logical, use Woodbury identity in beta update}

\item{diagX}{Logical, assume diagonal X}

\item{hyper_params}{List of hyperparameters (s_a, r_a, s_b, r_b, sigma_a, sigma_b)}
}
\value{
A list containing posterior samples matrices and a diagnostics summary table.
}
\description{
Fully Gibbs Sampler Function with Miller's Approximation
}
