% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_a_b.R
\name{Gibbs_a}
\alias{Gibbs_a}
\title{Sample posterior for hyperparameter 'a'}
\usage{
Gibbs_a(
  nu,
  a,
  prior = "gamma",
  s_a = 1.5,
  r_a = 1,
  scale_a = 1,
  method = "mh",
  mh_step = 0.05
)
}
\arguments{
\item{nu}{p*1 vector of local shrinkage parameters}

\item{a}{scalar, current value of 'a'}

\item{prior}{string, "gamma" or "hcauchy"}

\item{s_a}{scalar, shape parameter for Gamma prior}

\item{r_a}{scalar, rate parameter for Gamma prior}

\item{scale_a}{scalar, scale parameter for Half-Cauchy prior}

\item{method}{string, "miller" or "mh" (only applicable if prior_option="gamma")}

\item{mh_step}{scalar, step size for MH uniform proposal}
}
\value{
scalar a_new
}
\description{
Sample posterior for hyperparameter 'a'
}
