% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_beta.R
\name{Gibbs_beta}
\alias{Gibbs_beta}
\title{draw posterior beta}
\usage{
Gibbs_beta(X, y, omega, sigmaSq, nu, lambda)
}
\arguments{
\item{X}{n*p design matrix}

\item{y}{n*1 response vector}

\item{omega}{coefficient, real number}

\item{nu}{p*1 coefficient vector}

\item{lambda}{p*1 coefficient vector}

\item{sigmasq}{coefficient, real number}
}
\value{
posterior vector beta
}
\description{
draw posterior beta
}
