% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_lambda.R
\name{Gibbs_lambda}
\alias{Gibbs_lambda}
\title{draw posterior lambda}
\usage{
Gibbs_lambda(b, p, omega, sigmaSq, beta, nu)
}
\arguments{
\item{b}{hyper-parameter, real number}

\item{p}{ncol(X)}

\item{omega}{coefficient, real number}

\item{beta}{p*1 coefficient vector}

\item{nu}{p*1 coefficient vector}

\item{sigmasq}{coefficient, real number}
}
\value{
posterior vector lambda
}
\description{
draw posterior lambda
}
