% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_competition.R
\name{initialize_adaptive}
\alias{initialize_adaptive}
\title{Performs the Data-Adaptive Initialization Strategy via Candidate Model Competition.}
\usage{
initialize_adaptive(
  X,
  y,
  omega_init_guess,
  sigmaSq_init_guess,
  iter_pre_opt = 100,
  iter_selection = 5000,
  candidates = list(horseshoe = list(a = 0.5, b = 0.5), normal_gamma = list(a = 0.5, b =
    5), studentt = list(a = 5, b = 0.5)),
  woodbury = TRUE,
  delta1 = 1e-06,
  delta2 = 0.001,
  delta3 = 0.001,
  window_size = 5,
  diagX = FALSE
)
}
\arguments{
\item{X, y}{Model matrix and response vector.}

\item{omega_init_guess, sigmaSq_init_guess}{Initial guesses for omega and sigmaSq.}

\item{iter_pre_opt}{Iterations for the pre-optimization EM stage for each candidate.}

\item{iter_selection}{Number of post-burn-in samples for model selection.}

\item{candidates}{Candidate models and their initial a,b values.}

\item{woodbury}{Logical, use Woodbury identity in beta updates.}

\item{delta1, delta2, delta3}{Convergence tolerances.}

\item{window_size}{Size of the convergence window.}

\item{diagX}{Logical, assume diagonal X.}
}
\value{
A list containing winning_params and winner_name.
}
\description{
Performs the Data-Adaptive Initialization Strategy via Candidate Model Competition.
}
